// packages/semantics/design-system/src/providers/CodaProvider.types.generated.ts
export interface components {
  schemas: {
// ...
    TableReference: {
      /**
       * @description ID of the table.
       * @example grid-pqRst-U
       */
      id: string;
      /**
       * @description The type of this resource.
       * @enum {string}
       */
      type: "table";
      tableType: components["schemas"]["TableType"];
      /**
       * Format: url
       * @description API link to the table.
       * @example https://coda.io/apis/v1/docs/AbCDeFGH/tables/grid-pqRst-U
       */
      href: string;
      /**
       * Format: url
       * @description Browser-friendly link to the table.
       * @example https://coda.io/d/_dAbCDeFGH/#Teams-and-Tasks_tpqRst-U
       */
      browserLink: string;
      /**
       * @description Name of the table.
       * @example Tasks
       */
      name: string;
      parent?: components["schemas"]["PageReference"];
    };
    /** @description Metadata about a table. */
    Table: {
      /**
       * @description ID of the table.
       * @example grid-pqRst-U
       */
      id: string;
      /**
       * @description The type of this resource.
       * @enum {string}
       */
      type: "table";
      tableType: components["schemas"]["TableType"];
      /**
       * Format: url
       * @description API link to the table.
       * @example https://coda.io/apis/v1/docs/AbCDeFGH/tables/grid-pqRst-U
       */
      href: string;
      /**
       * Format: url
       * @description Browser-friendly link to the table.
       * @example https://coda.io/d/_dAbCDeFGH/#Teams-and-Tasks_tpqRst-U
       */
      browserLink: string;
      /**
       * @description Name of the table.
       * @example Tasks
       */
      name: string;
      parent: components["schemas"]["PageReference"];
      parentTable?: components["schemas"]["TableReference"];
      displayColumn: components["schemas"]["ColumnReference"];
      /**
       * @description Total number of rows in the table.
       * @example 130
       */
      rowCount: number;
      /** @description Any sorts applied to the table. */
      sorts: components["schemas"]["Sort"][];
      layout: components["schemas"]["Layout"];
      filter?: unknown & components["schemas"]["FormulaDetail"];
      /**
       * Format: date-time
       * @description Timestamp for when the table was created.
       * @example 2018-04-11T00:18:57.946Z
       */
      createdAt: string;
      /**
       * Format: date-time
       * @description Timestamp for when the table was last modified.
       * @example 2018-04-11T00:18:57.946Z
       */
      updatedAt: string;
    };
    /** @description List of tables. */
    TableList: {
      items: components["schemas"]["TableReference"][];
      /**
       * Format: url
       * @description API link to these results
       * @example https://coda.io/apis/v1/docs/AbCDeFGH/tables?limit=20
       */
      href?: string;
      nextPageToken?: components["schemas"]["nextPageToken"];
      nextPageLink?: components["schemas"]["nextPageLink"] & string;
    };
// ...
    Column: {
      /**
       * @description ID of the column.
       * @example c-tuVwxYz
       */
      id: string;
      /**
       * @description The type of this resource.
       * @enum {string}
       */
      type: "column";
      /**
       * Format: url
       * @description API link to the column.
       * @example https://coda.io/apis/v1/docs/AbCDeFGH/tables/grid-pqRst-U/columns/c-tuVwxYz
       */
      href: string;
      /**
       * @description Name of the column.
       * @example Completed
       */
      name: string;
      /**
       * @description Whether the column is the display column.
       * @example true
       */
      display?: boolean;
      /**
       * @description Whether the column has a formula set on it.
       * @example true
       */
      calculated?: boolean;
      /**
       * @description Formula on the column.
       * @example thisRow.Created()
       */
      formula?: string;
      /**
       * @description Default value formula for the column.
       * @example Test
       */
      defaultValue?: string;
      format: components["schemas"]["ColumnFormat"];
    };
// ...
    ColumnFormat:
      | components["schemas"]["ButtonColumnFormat"]
      | components["schemas"]["CheckboxColumnFormat"]
      | components["schemas"]["DateColumnFormat"]
      | components["schemas"]["DateTimeColumnFormat"]
      | components["schemas"]["DurationColumnFormat"]
      | components["schemas"]["EmailColumnFormat"]
      | components["schemas"]["LinkColumnFormat"]
      | components["schemas"]["CurrencyColumnFormat"]
      | components["schemas"]["ImageReferenceColumnFormat"]
      | components["schemas"]["NumericColumnFormat"]
      | components["schemas"]["ReferenceColumnFormat"]
      | components["schemas"]["SelectColumnFormat"]
      | components["schemas"]["SimpleColumnFormat"]
      | components["schemas"]["ScaleColumnFormat"]
      | components["schemas"]["SliderColumnFormat"]
      | components["schemas"]["TimeColumnFormat"];
// ...
    ColumnList: {
      items: components["schemas"]["Column"][];
      /**
       * Format: url
       * @description API link to these results
       * @example https://coda.io/apis/v1/docs/AbCDeFGH/tables/grid-pqRst-U/columns?limit=20
       */
      href?: string;
      nextPageToken?: components["schemas"]["nextPageToken"];
      nextPageLink?: components["schemas"]["nextPageLink"] & string;
    };
    /** @description Info about a row. */
    Row: {
      /**
       * @description ID of the row.
       * @example i-tuVwxYz
       */
      id: string;
      /**
       * @description The type of this resource.
       * @enum {string}
       */
      type: "row";
      /**
       * Format: url
       * @description API link to the row.
       * @example https://coda.io/apis/v1/docs/AbCDeFGH/tables/grid-pqRst-U/rows/i-RstUv-W
       */
      href: string;
      /**
       * @description The display name of the row, based on its identifying column.
       * @example Apple
       */
      name: string;
      /**
       * @description Index of the row within the table.
       * @example 7
       */
      index: number;
      /**
       * Format: url
       * @description Browser-friendly link to the row.
       * @example https://coda.io/d/_dAbCDeFGH#Teams-and-Tasks_tpqRst-U/_rui-tuVwxYz
       */
      browserLink: string;
      /**
       * Format: date-time
       * @description Timestamp for when the row was created.
       * @example 2018-04-11T00:18:57.946Z
       */
      createdAt: string;
      /**
       * Format: date-time
       * @description Timestamp for when the row was last modified.
       * @example 2018-04-11T00:18:57.946Z
       */
      updatedAt: string;
      /**
       * @description Values for a specific row, represented as a hash of column IDs (or names with `useColumnNames`) to values.
       * @example {
       *       "c-tuVwxYz": "Apple",
       *       "c-bCdeFgh": [
       *         "$12.34",
       *         "$56.78"
       *       ]
       *     }
       */
      values: {
        [key: string]: components["schemas"]["CellValue"];
      };
    };
// ...
    RowList: {
      items: components["schemas"]["Row"][];
      /**
       * Format: url
       * @description API link to these results
       * @example https://coda.io/apis/v1/docs/AbCDeFGH/tables/grid-pqRst-U/rows?limit=20
       */
      href?: string;
      nextPageToken?: components["schemas"]["nextPageToken"];
      nextPageLink?: components["schemas"]["nextPageLink"] & string;
      nextSyncToken?: components["schemas"]["nextSyncToken"];
    };
// ...
    CellValue:
      | components["schemas"]["Value"]
      | components["schemas"]["RichValue"];
// ...
    RowUpdateResult: components["schemas"]["DocumentMutateResponse"] & {
      /**
       * @description ID of the updated row.
       * @example i-tuVwxYz
       */
      id: string;
    };
// ...
    MutationStatus: {
      /**
       * @description Returns whether the mutation has completed.
       * @example true
       */
      completed: boolean;
      /**
       * @description A warning if the mutation completed but with caveats.
       * @example Initial page HTML was invalid.
       */
      warning?: string;
    };
// ...
  };
// ...
}